requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
databases[asyncpg]
pydantic
alembic
psycopg2-binary
celery[redis]
redis
python-dotenv
